requirements.txt
pytest
Appium-Python-Client~=2.7.0
selenium~=4.5.0
cryptography
python-dotenv
pytest_check
black
mypy
isort
flake8
pre-commit
allure-pytest
webdriver-manager
loguru
pytest-html
pytest-xdist
interrogate
pytest-rerunfailures
allure-combine
requests
types-requests
python-dateutil
